+++ src/main.cpp
@@ -3,6 +3,7 @@
 
 int main()
 {
+    
     FreiaUI ui;
     ClientConnect client;